backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <mylesngamez@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9"
fastapi = "*"
openai = "*"
python-dotenv = "*"
sqlalchemy = "*"
asyncpg = "*"
greenlet = "*"
psycopg2 = "*"
uvicorn = "*"
websockets = "^15.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
